add body argument to Do()
In some cases, such as when handling digest auth,
we might need to retry the request.  We shouldn't
be using the original request object however
because the Body will have been read by the first
request.  To fix this a new argument has been added
to Do() so that the original body can be passed along
to the next request.

Note, the original body could be copied.  But this seemed
like a better implementation because it saves a couple
of allocations and potentially memory too depending on
the body size.