_account_id should be omitted if empty (closes #16)
The AccountInfo struct is used in other structs, such
as ApprovalInfo, and _account_id may not always be
set in those cases. In these cases if the _account_id
field is not setup with 'omitempty' then the field will
be included anyway even if Gerrit does not return it.

Thanks to @perolausson for finding and explaining the
issue.