fix type of EventInfo.PatchSet.Number
Depending on the version of the events-log plugin
and the version of Gerrit you can end up in a situation
where some values of EventInfo.PatchSet.Number are a
string and other are an integer. This change works around
the problem by replacing the Number field with a custom
type that can handle either strings or integers.

Note, this does not change the underlying type of the
Number field.