fix NewClient() url parsing for passwords containing / or other url characters
Depending on the contents of the username and password the default
url.Parse may not work. The below is an example URL that
would end up being parsed incorrectly with url.Parse:
  http://admin:ZOSOKjgV/kgEkN0bzPJp+oGeJLqpXykqWFJpon/Ckg@localhost:38607

So instead of depending on url.Parse we'll try using a regular expression
first to match a specific pattern instead. This won't be 100% reliable
in all cases which is why we fall back onto using url.Parse anyway.