Add Avatars field to AccountInfo.
This change implements support for the avatars plugin.
It appears to be available on many Gerrit instances hosted by Google.

Unfortunately, I have not been able to find any API documentation or
source for this augmentation of the AccountInfo entity. This change
is implemented based on looking at real responses from Google-hosted
Gerrit instances. E.g.:

	$ curl 'https://gerrit-review.googlesource.com/accounts/1010008'
	)]}'
	{
	  "_account_id": 1010008,
	  "name": "Example Example",
	  "email": "example@example.com",
	  "avatars": [
	    {
	      "url": "https://lh3.googleusercontent.com/-qjpKeQViQbI/AAAAAAAAAAI/AAAAAAAAAAA/_CVwpoi1Y8Y/s26-p/photo.jpg",
	      "height": 26
	    },
	    {
	      "url": "https://lh3.googleusercontent.com/-qjpKeQViQbI/AAAAAAAAAAI/AAAAAAAAAAA/_CVwpoi1Y8Y/s32-p/photo.jpg",
	      "height": 32
	    },
	    {
	      "url": "https://lh3.googleusercontent.com/-qjpKeQViQbI/AAAAAAAAAAI/AAAAAAAAAAA/_CVwpoi1Y8Y/s100-p/photo.jpg",
	      "height": 100
	    }
	  ]
	}

Resolves #51.