Expose the client's BaseURL.
The motivation for doing this is as follows.

NewClient takes a gerritURL string, which is then parsed into a baseURL
internal to the gerrit.Client.

Some users of gerrit.Client may need to know the Gerrit instance that
the client targets. Currently, they'd have to be provided the target
separately, which can lead to a mismatch from the actual gerrit.Client
Gerrit instance:

	client, err := gerrit.NewClient("https://go-review.googlesource.com/", nil)
	if err != nil {
		return err
	}
	service := gerritapi.NewService(client, "https://upspin-review.googlesource.com/") // mismatch

BaseURL method is added to allow users to find out the Gerrit URL
endpoint directly from the gerrit.Client, removing the chance for
any mismatches:

	client, err := gerrit.NewClient("https://go-review.googlesource.com/", nil)
	if err != nil {
		return err
	}
	service := gerritapi.NewService(client) // no mismatch possible, gerritapi queries actual client.BaseURL()