changes: fill in the rest of Changes.ChangeInput
The ChangeInput type was recently added with just the bare-bones
fields in place. This commit adds the rest of the fields per the
current gerrit documentation.

Updated TestChangesService_CreateChange to have stronger assertions
about empty values (recursively checks nested structures), and added
test cases specific to exercise ChangeInput.Merge and
ChangeInput.Author.

This change is a result of the discussion on PR #96.