Fix handling of Skip and Start parameters
The way we treated Skip and Start parameters was inconsistent and wrong.
In some calls we send both parameters (s, start, S). This causes
and undefined behaviour at Gerrit itself.

Additional the Skip/Start parameters are numbers and not strings.
Here we fixed the Go types.

Thanks a lot to @michaeldorner for his initial work in #48

Related:
- https://github.com/andygrunwald/go-gerrit/pull/48