Add golang context support
This is a large cross cutting change to the existing interfaces for
go-gerrit that introduces context support to all interfaces drilling
down to the HTTP level.

Contexts enable support for both passing information through (eg:
tracing, logging, etc) and features like cancellation of slow
operations (eg: REST calls).

In order for consumers of this breaking change to adapt their code to
use the new interface they must add a context parameter to their
invocations.  In most cases this can just be `context.Background()`