fix: handle changes/<change>/patch JSON marshalling error
The changes/<change>/revisions/<revision>/patch endpoint in gerrit
returns a base64 encoded diff by default, which cannot be marshalled as
JSON. Creating a unit test for the patch endpoint results in the 'invalid
character <some random char> looking for beginning of value` error (same
as https://github.com/andygrunwald/go-gerrit/issues/166).

This change fixes the issue by providing an input to Client.Do() which
implements the Writer interface to avoid the JSON marshalling by
Client.Do() and allow the caller to handle the response appropriately.