don't treat + as a special character for queries
This special behavior for the "+" character was added as part of PR #22
to resolve issue #18 around 9 years ago, but the issue described there
seems more like a problem caused by "+" being used instead of " "
(a space) in the original search query, and we haven't received new
information to suggest otherwise. So, undo the special behavior for the
"+" character.

GitHub-Pull-Request: https://github.com/andygrunwald/go-gerrit/pull/193